refactor: move rlp_utils to block_access_lists; bal -> block_access_lists
