+++ src/ethereum/forks/amsterdam/blocks.py
@@ -18,8 +18,8 @@ from ethereum_types.numeric import U64, U256, Uint
 
 from ethereum.crypto.hash import Hash32
 
+from .block_access_lists.rlp_types import BlockAccessList
 from .fork_types import Address, Bloom, Root
-from .rlp_types import BlockAccessList
 from .transactions import (
     AccessListTransaction,
     BlobTransaction,
@@ -242,7 +242,7 @@ class Header:
     [SHA2-256]: https://en.wikipedia.org/wiki/SHA-2
     """
 
-    bal_hash: Hash32
+    block_access_list_hash: Hash32
     """
     [SHA2-256] hash of the Block Access List containing all accounts and
     storage locations accessed during block execution. Introduced in
