+++ src/ethereum/genesis.py
@@ -259,8 +259,8 @@ def add_genesis_block(
     if has_field(hardfork.Header, "requests_hash"):
         fields["requests_hash"] = Hash32(b"\0" * 32)
 
-    if has_field(hardfork.Header, "bal_hash"):
-        fields["bal_hash"] = Hash32(b"\0" * 32)
+    if has_field(hardfork.Header, "block_access_list_hash"):
+        fields["block_access_list_hash"] = Hash32(b"\0" * 32)
 
     genesis_header = hardfork.Header(**fields)
