+++ src/ethereum_spec_tools/evm_tools/loaders/fork_loader.py
@@ -125,7 +125,7 @@ class ForkLoad:
         return self._module("transactions").signing_hash_155
 
     @property
-    def build(self) -> Any:
+    def build_block_access_list(self) -> Any:
         """build function of the fork"""
         return self._module("block_access_lists").build_block_access_list
