+++ src/ethereum_spec_tools/evm_tools/t8n/__init__.py
@@ -316,9 +316,9 @@ class T8N(Load):
             self.fork.process_general_purpose_requests(block_env, block_output)
 
         if self.fork.is_after_fork("ethereum.forks.amsterdam"):
-                block_output.block_access_list = self.fork.build(
-                    block_env.state.change_tracker.block_access_list_builder
-                )
+            block_output.block_access_list = self.fork.build_block_access_list(
+                block_env.state.change_tracker.block_access_list_builder
+            )
 
     def run_blockchain_test(self) -> None:
         """
