+++ tests/amsterdam/test_rlp.py
@@ -4,7 +4,7 @@ from ethereum_types.bytes import Bytes, Bytes0, Bytes8, Bytes32
 from ethereum_types.numeric import U64, U256, Uint
 
 from ethereum.forks.amsterdam.blocks import Block, Header, Log, Receipt, Withdrawal
-from ethereum.forks.amsterdam.rlp_types import BlockAccessList
+from ethereum.forks.amsterdam.block_access_lists.rlp_types import BlockAccessList
 from ethereum.forks.amsterdam.transactions import (
     Access,
     AccessListTransaction,
@@ -114,7 +114,7 @@ header = Header(
     blob_gas_used=U64(7),
     excess_blob_gas=U64(8),
     requests_hash=hash7,
-    bal_hash=hash1,  # Added missing bal_hash
+    block_access_list_hash=hash1,  # Added missing block_access_list_hash
 )
 
 block = Block(
